backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
python-dotenv==1.0.0
email-validator>=2.0.0
